feat: Created a separate DeletionEvent
