+++ engine/src/main/java/org/destinationsol/removal/events/DeletionEvent.java
@@ -15,18 +15,15 @@
  */
 package org.destinationsol.removal.events;
 
-import com.badlogic.gdx.physics.box2d.Body;
-import org.destinationsol.removal.components.SlatedForDeletion;
 import org.destinationsol.removal.systems.DeletionUpdateSystem;
 import org.destinationsol.removal.systems.DestructionSystem;
 import org.terasology.gestalt.entitysystem.event.Event;
 
 /**
- * Event that indicates that an entity no longer needs to exist and thus can be removed. Entities should not be deleted
- * when this event is processed. Instead, a {@link SlatedForDeletion} component should be added to the entity. Every
- * tick, the {@link DeletionUpdateSystem} sends a {@link DeletionEvent} to each entity with a SlatedForDeletion
- * component, which causes the entity to actually be deleted by the {@link DestructionSystem}. This is done so that
- * certain resources, such as libGDX's {@link Body} class, can be disposed of properly before the entity is deleted.
+ * Every tick, the {@link DeletionUpdateSystem} sends a DeletionEvent to each entity with a SlatedForDeletion component,
+ * which causes the entity to be deleted by the {@link DestructionSystem}. This is different than a
+ * {@link ShouldBeDestroyedEvent} - that is for systems to respond to the destruction of an entity before it happens,
+ * e.g. to prevent it from being destroyed. This event is the actual deletion of the entity.
  */
-public class RemovalForOptimizationEvent implements Event {
+public class DeletionEvent implements Event {
 }
